refactor: refactor cjs to esm
